Nov enhancements valar

+++ database/migrations/2023_07_28_103416_create_total_management_payroll_attachments_table.php
@@ -42,6 +42,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('total_management_payroll_attachments');
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::dropIfExists('total_management_payroll_attachments');
+        }
     }
 };
