+++ database/migrations/2023_08_07_080055_create_payroll_bulk_upload_table.php
@@ -38,6 +38,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('payroll_bulk_upload');
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::dropIfExists('payroll_bulk_upload');
+        }
     }
 };
